[g8r] Add gatesim benchmark
