+++ xlsynth-g8r/Cargo.toml
@@ -18,6 +18,7 @@ tempfile = "3.3"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 flate2 = { version = "1.1", features = ["zlib-rs"], default-features = false }
+half = "2.6"
 
 [dev-dependencies]
 test-case = "3.3.1"
@@ -26,8 +27,8 @@ libfuzzer-sys = "0.4"
 pretty_assertions = "1.4.1"
 regex = "1.5"
 cargo_metadata = "0.19.2"
-half = "2.6"
 rand = "0.8"
+criterion = "0.5"
 
 [lib]
 name = "xlsynth_g8r"
@@ -40,3 +41,7 @@ path = "src/main.rs"
 [[bin]]
 name = "try-parse-liberty"
 path = "src/liberty/try_parse.rs"
+
+[[bench]]
+name = "bf16_mul_gatesim_bench"
+harness = false
